+++ doc/version-history.rst
@@ -13,6 +13,7 @@ Version History
 v0.4.2
 ======
 * Pass logs to CSC logs for debugging.
+* Start the tunnel even when using open_tunnel.
 
 v0.4.1
 ======